requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
claude-agent-sdk>=0.1.0
python-dotenv==1.0.1
pydantic==2.9.2
bcrypt==4.2.0
itsdangerous==2.2.0
